LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-path-animation-lL-vV-hH-inverse.tentative-expected.txt


FAIL Path animation where coordinate modes of start and end differ (l-L, v-V and h-H) assert_equals: expected "M 15.00 15.00 L -15.00 0.00 V -15.00 H 0.00 Z" but got "M 15.00 15.00 l -30.00 -15.00 v -15.00 h 15.00 Z"
